refactor(core): IWopiHostExtensions + builder seams (#409 5.4/5.5)
5.4: replace static WopiExtensions.GetWopiCheckFileInfo /
GetWopiCheckContainerInfo / BuildCheckFolderInfo extension methods —
which pulled deps from HttpContext.RequestServices — with constructor-
injected ICheckFileInfoBuilder / ICheckContainerInfoBuilder /
ICheckFolderInfoBuilder. Default impls in WopiHost.Core; File and
Container builders fire the configured IWopiHostExtensions hook
internally, Folder builder stays sync to preserve the #363 Infer# fix.

5.5: remove the 6 behavior delegates (OnCheckFileInfo /
OnCheckContainerInfo / OnCheckFolderInfo / OnCheckEcosystem / OnPutFile /
OnPutRelativeFile) from WopiHostOptions. They move to a single
IWopiHostExtensions interface with virtual methods (now also taking a
CancellationToken); hosts extend the supplied WopiHostExtensions base
class and override only what they need. WopiHostOptions is now pure
data.

Context records (WopiCheck*Context, WopiPut*Context) move from
WopiHost.Core.Models to WopiHost.Abstractions so the new interfaces can
live alongside the rest of the contracts.

Breaking change for hosts that customized callbacks via Configure<
WopiHostOptions>. Migration: subclass WopiHostExtensions and register
the subclass on IWopiHostExtensions. Validator sample updated as a
worked example.
