+++ src/WopiHost.Abstractions/README.md
@@ -21,7 +21,9 @@ dotnet add package WopiHost.Abstractions
 | [`IWopiLockComparer`](IWopiLockComparer.cs) | Compares two lock ids for equivalence — used by every controller validation and by both providers' atomic compare-and-swap. | `OrdinalWopiLockComparer` (default, byte-exact), `JsonShapedWopiLockComparer` (opt-in, OOS-tolerant) | Replace only if you observe a specific WOPI client mutating lock ids between round-trips. |
 | [`IWopiPermissionProvider`](IWopiPermissionProvider.cs) | What a `ClaimsPrincipal` can do with a file/container. | `DefaultWopiPermissionProvider` (reads from token claims) | Almost always — this is where your ACL model plugs in. |
 | [`IWopiAccessTokenService`](IWopiAccessTokenService.cs) | Issue and validate access tokens. | `JwtAccessTokenService` (signed JWT) | Rarely — only if you need opaque/revocable tokens or external issuance. |
-| [`IWopiHostCapabilities`](IWopiHostCapabilities.cs) | Feature flags reported in `CheckFileInfo`. | `WopiHostCapabilities` | Override individual flags via `WopiHostOptions.OnCheckFileInfo`. |
+| [`IWopiHostCapabilities`](IWopiHostCapabilities.cs) | Feature flags reported in `CheckFileInfo`. | `WopiHostCapabilities` | Override individual flags by extending [`WopiHostExtensions`](WopiHostExtensions.cs) and overriding `OnCheckFileInfoAsync`. |
+| [`IWopiHostExtensions`](IWopiHostExtensions.cs) | Single host-customization seam for `CheckFileInfo` / `CheckContainerInfo` / `CheckFolderInfo` / `CheckEcosystem` / `PutFile` / `PutRelativeFile`. | `WopiHostExtensions` (pass-through) | Subclass when you want to layer custom properties, rewrite URLs, or hook write-completion telemetry. |
+| [`ICheckFileInfoBuilder`](ICheckFileInfoBuilder.cs) / [`ICheckContainerInfoBuilder`](ICheckContainerInfoBuilder.cs) / [`ICheckFolderInfoBuilder`](ICheckFolderInfoBuilder.cs) | Build the corresponding response shapes. The default implementations fire the matching `IWopiHostExtensions` hook. | `DefaultCheck*InfoBuilder` (in `WopiHost.Core`) | Replace when you need scoped services that don't fit in the host-extensions seam, or to short-circuit the default population entirely. |
 
 The defaults ship in `WopiHost.Core`, `WopiHost.FileSystemProvider`, `WopiHost.MemoryLockProvider`, `WopiHost.AzureStorageProvider`, and `WopiHost.AzureLockProvider` — selected via `WopiHostOptions.StorageProviderAssemblyName` / `LockProviderAssemblyName`.
