+++ src/WopiHost.Core/Models/WopiHostOptions.cs
@@ -1,4 +1,4 @@
-﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations;
 using Microsoft.AspNetCore.Http;
 using WopiHost.Abstractions;
 using WopiHost.Discovery;
@@ -6,7 +6,10 @@ using WopiHost.Discovery;
 namespace WopiHost.Core.Models;
 
 /// <summary>
-/// Configuration class for WopiHost.Core
+/// Configuration class for WopiHost.Core. Pure data — host-customization behavior lives on
+/// <see cref="IWopiHostExtensions"/>; replace per-resource response shaping by registering
+/// custom <see cref="ICheckFileInfoBuilder"/> / <see cref="ICheckContainerInfoBuilder"/> /
+/// <see cref="ICheckFolderInfoBuilder"/> implementations.
 /// </summary>
 public class WopiHostOptions : IDiscoveryOptions
 {
@@ -31,49 +34,6 @@ public class WopiHostOptions : IDiscoveryOptions
     /// Name of the assembly containing the implementation of the <see cref="Abstractions.IWopiLockProvider"/>
     /// </summary>
     public string? LockProviderAssemblyName { get; set; }
-
-    /// <summary>
-    /// Callback for the CheckFileInfo operation.
-    /// </summary>
-    public Func<WopiCheckFileInfoContext, Task<WopiCheckFileInfo>> OnCheckFileInfo { get; set; } = c => Task.FromResult(c.CheckFileInfo);
-
-    /// <summary>
-    /// Callback for the CheckContainerInfo operation.
-    /// </summary>
-    public Func<WopiCheckContainerInfoContext, Task<WopiCheckContainerInfo>> OnCheckContainerInfo { get; set; } = c => Task.FromResult(c.CheckContainerInfo);
-
-    /// <summary>
-    /// Callback for the CheckFolderInfo operation.
-    /// </summary>
-    public Func<WopiCheckFolderInfoContext, Task<WopiCheckFolderInfo>> OnCheckFolderInfo { get; set; } = c => Task.FromResult(c.CheckFolderInfo);
-
-    /// <summary>
-    /// Callback for the CheckEcosystem operation.
-    /// </summary>
-    public Func<WopiCheckEcosystemContext, Task<WopiCheckEcosystem>> OnCheckEcosystem { get; set; } = c => Task.FromResult(c.CheckEcosystem);
-
-    /// <summary>
-    /// Callback raised after a successful <c>PutFile</c> write. Receives the updated file plus
-    /// the user ids parsed from the optional <c>X-WOPI-Editors</c> request header. Use this for
-    /// audit trails and last-touch metadata; the default is a no-op.
-    /// </summary>
-    /// <remarks>
-    /// Throwing inside this callback turns the response into a 500. For best-effort bookkeeping,
-    /// catch exceptions inside the handler.
-    /// </remarks>
-    public Func<WopiPutFileContext, Task> OnPutFile { get; set; } = _ => Task.CompletedTask;
-
-    /// <summary>
-    /// Callback raised after a successful <c>PutRelativeFile</c> write. Surfaces the optional
-    /// <c>X-WOPI-FileConversion</c> (presence flag) and <c>X-WOPI-Size</c> headers alongside
-    /// both the original and newly-created files. Hosts that want to flag conversion-context
-    /// uploads or record declared-size telemetry plug in here; the default is a no-op.
-    /// </summary>
-    /// <remarks>
-    /// Throwing inside this callback turns the response into a 500. For best-effort bookkeeping,
-    /// catch exceptions inside the handler.
-    /// </remarks>
-    public Func<WopiPutRelativeFileContext, Task> OnPutRelativeFile { get; set; } = _ => Task.CompletedTask;
 
     /// <summary>
     /// Value written to the <c>X-WOPI-Lock</c> response header when the file is unlocked but the
