+++ test/WopiHost.Core.Tests/Infrastructure/DefaultCheckInfoBuilderTests.cs
@@ -1,0 +1,505 @@
+using System.Security.Claims;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Routing;
+using Moq;
+using WopiHost.Abstractions;
+using WopiHost.Core.Infrastructure;
+
+namespace WopiHost.Core.Tests.Infrastructure;
+
+/// <summary>
+/// Tests for the default <see cref="ICheckFileInfoBuilder"/>, <see cref="ICheckContainerInfoBuilder"/>,
+/// and <see cref="ICheckFolderInfoBuilder"/> implementations. Ported from the now-deleted
+/// <c>WopiExtensionsTests</c> when the <c>GetWopiCheckFileInfo</c> / <c>GetWopiCheckContainerInfo</c> /
+/// <c>BuildCheckFolderInfo</c> extension methods were replaced by these builders.
+/// </summary>
+public class DefaultCheckInfoBuilderTests
+{
+    private static Mock<IWopiPermissionProvider> CreatePermissionProvider()
+    {
+        var permissionProvider = new Mock<IWopiPermissionProvider>();
+        permissionProvider
+            .Setup(_ => _.GetFilePermissionsAsync(It.IsAny<ClaimsPrincipal>(), It.IsAny<IWopiFile>(), It.IsAny<CancellationToken>()))
+            .ReturnsAsync(WopiFilePermissions.None);
+        permissionProvider
+            .Setup(_ => _.GetContainerPermissionsAsync(It.IsAny<ClaimsPrincipal>(), It.IsAny<IWopiFolder>(), It.IsAny<CancellationToken>()))
+            .ReturnsAsync(WopiContainerPermissions.None);
+        return permissionProvider;
+    }
+
+    [Fact]
+    public async Task GetWopiCheckFileInfo_ReturnsCorrectInfo()
+    {
+        var mockFile = new Mock<IWopiFile>();
+        // Stub Checksum so GetEncodedSha256 takes the early-return path; otherwise it would
+        // call OpenReadAsync (also unmocked), and ComputeHashAsync(null) throws.
+        mockFile.Setup(f => f.Checksum).Returns(new ReadOnlyMemory<byte>([0]));
+        mockFile.Setup(f => f.Name).Returns("test");
+        mockFile.Setup(f => f.Owner).Returns("owner");
+        mockFile.Setup(f => f.Extension).Returns("txt");
+        mockFile.Setup(f => f.LastWriteTimeUtc).Returns(DateTime.UtcNow);
+        mockFile.Setup(f => f.Exists).Returns(true);
+        mockFile.Setup(f => f.Length).Returns(12345);
+
+        var capabilities = new WopiHostCapabilities
+        {
+            SupportsCoauth = true,
+            SupportsFolders = true,
+            SupportsLocks = true,
+            SupportsGetLock = true,
+            SupportsExtendedLockLength = true,
+            SupportsEcosystem = true,
+            SupportsGetFileWopiSrc = true,
+            SupportedShareUrlTypes = ["ReadOnly", "ReadWrite"],
+            SupportsScenarioLinks = true,
+            SupportsSecureStore = true,
+            SupportsUpdate = true,
+            SupportsCobalt = true,
+            SupportsRename = true,
+            SupportsDeleteFile = true,
+            SupportsUserInfo = true,
+            SupportsFileCreation = true
+        };
+        var httpContext = new DefaultHttpContext();
+
+        var builder = new DefaultCheckFileInfoBuilder(CreatePermissionProvider().Object, new WopiHostExtensions());
+        var result = await builder.BuildAsync(mockFile.Object, httpContext, capabilities, "userInfo text");
+
+        Assert.Equal("owner", result.OwnerId);
+        Assert.Equal("test.txt", result.BaseFileName);
+        Assert.Equal(".txt", result.FileExtension);
+        Assert.Equal(0, result.FileNameMaxLength);
+        Assert.Equal(12345, result.Size);
+        Assert.Equal("userInfo text", result.UserInfo);
+        Assert.True(result.SupportsCoauth);
+        Assert.True(result.SupportsFolders);
+        Assert.True(result.SupportsLocks);
+        Assert.True(result.SupportsGetLock);
+        Assert.True(result.SupportsExtendedLockLength);
+        Assert.True(result.SupportsEcosystem);
+        Assert.True(result.SupportsGetFileWopiSrc);
+        Assert.Equal(["ReadOnly", "ReadWrite"], result.SupportedShareUrlTypes);
+        Assert.True(result.SupportsScenarioLinks);
+        Assert.True(result.SupportsSecureStore);
+        Assert.True(result.SupportsUpdate);
+        Assert.True(result.SupportsCobalt);
+        Assert.True(result.SupportsRename);
+        Assert.True(result.SupportsDeleteFile);
+        Assert.True(result.SupportsUserInfo);
+        Assert.True(result.SupportsFileCreation);
+        Assert.True(result.IsAnonymousUser);
+    }
+
+    [Fact]
+    public async Task GetWopiCheckFileInfo_WithWritableStorageProvider_ReturnsCorrectInfo()
+    {
+        var mockFile = new Mock<IWopiFile>();
+        // Stub Checksum so GetEncodedSha256 takes the early-return path; otherwise it would
+        // call OpenReadAsync (also unmocked), and ComputeHashAsync(null) throws.
+        mockFile.Setup(f => f.Checksum).Returns(new ReadOnlyMemory<byte>([0]));
+        mockFile.Setup(f => f.Name).Returns("test");
+        mockFile.Setup(f => f.Owner).Returns("owner");
+        mockFile.Setup(f => f.Extension).Returns("txt");
+        mockFile.Setup(f => f.LastWriteTimeUtc).Returns(DateTime.UtcNow);
+        mockFile.Setup(f => f.Exists).Returns(true);
+        mockFile.Setup(f => f.Length).Returns(12345);
+
+        var writableStorageProvider = new Mock<IWopiWritableStorageProvider>();
+        writableStorageProvider.Setup(_ => _.FileNameMaxLength).Returns(13);
+        var httpContext = new DefaultHttpContext();
+
+        var builder = new DefaultCheckFileInfoBuilder(
+            CreatePermissionProvider().Object,
+            new WopiHostExtensions(),
+            writableStorageProvider.Object);
+        var result = await builder.BuildAsync(mockFile.Object, httpContext);
+
+        Assert.Equal(13, result.FileNameMaxLength);
+    }
+
+    private sealed class StubLinkGenerator(string returnValue) : LinkGenerator
+    {
+        public override string? GetPathByAddress<TAddress>(HttpContext httpContext, TAddress address, RouteValueDictionary values, RouteValueDictionary? ambientValues = null, PathString? pathBase = null, FragmentString fragment = default, LinkOptions? options = null) => returnValue;
+        public override string? GetPathByAddress<TAddress>(TAddress address, RouteValueDictionary values, PathString pathBase = default, FragmentString fragment = default, LinkOptions? options = null) => returnValue;
+        public override string? GetUriByAddress<TAddress>(HttpContext httpContext, TAddress address, RouteValueDictionary values, RouteValueDictionary? ambientValues = null, string? scheme = null, HostString? host = null, PathString? pathBase = null, FragmentString fragment = default, LinkOptions? options = null) => returnValue;
+        public override string? GetUriByAddress<TAddress>(TAddress address, RouteValueDictionary values, string? scheme, HostString host, PathString pathBase = default, FragmentString fragment = default, LinkOptions? options = null) => returnValue;
+    }
+
+    [Fact]
+    public async Task GetWopiCheckFileInfo_PopulatesFileUrl_WhenLinkGeneratorIsRegistered()
+    {
+        var mockFile = new Mock<IWopiFile>();
+        // Stub Checksum so GetEncodedSha256 takes the early-return path; otherwise it would
+        // call OpenReadAsync (also unmocked), and ComputeHashAsync(null) throws.
+        mockFile.Setup(f => f.Checksum).Returns(new ReadOnlyMemory<byte>([0]));
+        mockFile.Setup(f => f.Name).Returns("test");
+        mockFile.Setup(f => f.Owner).Returns("owner");
+        mockFile.Setup(f => f.Extension).Returns("txt");
+        mockFile.Setup(f => f.Identifier).Returns("WOPITEST");
+        mockFile.Setup(f => f.LastWriteTimeUtc).Returns(DateTime.UtcNow);
+
+        const string expected = "https://localhost/wopi/files/WOPITEST/contents?access_token=tok";
+        var linkGenerator = new StubLinkGenerator(expected);
+
+        var httpContext = new DefaultHttpContext();
+        httpContext.Request.Scheme = "https";
+        httpContext.Request.Host = new HostString("localhost");
+
+        var builder = new DefaultCheckFileInfoBuilder(
+            CreatePermissionProvider().Object,
+            new WopiHostExtensions(),
+            writableStorageProvider: null,
+            linkGenerator: linkGenerator);
+        var result = await builder.BuildAsync(mockFile.Object, httpContext);
+
+        Assert.NotNull(result.FileUrl);
+        Assert.Equal(expected, result.FileUrl.ToString());
+    }
+
+    [Fact]
+    public async Task GetWopiCheckFileInfo_OverridesFileUrl_WhenOnCheckFileInfoSetsIt()
+    {
+        var mockFile = new Mock<IWopiFile>();
+        // Stub Checksum so GetEncodedSha256 takes the early-return path; otherwise it would
+        // call OpenReadAsync (also unmocked), and ComputeHashAsync(null) throws.
+        mockFile.Setup(f => f.Checksum).Returns(new ReadOnlyMemory<byte>([0]));
+        mockFile.Setup(f => f.Name).Returns("test");
+        mockFile.Setup(f => f.Owner).Returns("owner");
+        mockFile.Setup(f => f.Extension).Returns("txt");
+        mockFile.Setup(f => f.Identifier).Returns("WOPITEST");
+        mockFile.Setup(f => f.LastWriteTimeUtc).Returns(DateTime.UtcNow);
+
+        var linkGenerator = new StubLinkGenerator("https://localhost/wopi/files/WOPITEST/contents?access_token=tok");
+
+        var cdnUrl = new Uri("https://cdn.example.com/file");
+        var extensions = new RewritingExtensions
+        {
+            CheckFileInfoHandler = (context, _) => { context.CheckFileInfo.FileUrl = cdnUrl; return Task.FromResult(context.CheckFileInfo); },
+        };
+
+        var httpContext = new DefaultHttpContext();
+        httpContext.Request.Scheme = "https";
+        httpContext.Request.Host = new HostString("localhost");
+
+        var builder = new DefaultCheckFileInfoBuilder(
+            CreatePermissionProvider().Object,
+            extensions,
+            writableStorageProvider: null,
+            linkGenerator: linkGenerator);
+        var result = await builder.BuildAsync(mockFile.Object, httpContext);
+
+        Assert.Equal(cdnUrl, result.FileUrl);
+    }
+
+    [Fact]
+    public async Task GetWopiCheckFileInfo_LeavesFileUrlNull_WhenLinkGeneratorIsMissing()
+    {
+        var mockFile = new Mock<IWopiFile>();
+        // Stub Checksum so GetEncodedSha256 takes the early-return path; otherwise it would
+        // call OpenReadAsync (also unmocked), and ComputeHashAsync(null) throws.
+        mockFile.Setup(f => f.Checksum).Returns(new ReadOnlyMemory<byte>([0]));
+        mockFile.Setup(f => f.Name).Returns("test");
+        mockFile.Setup(f => f.Owner).Returns("owner");
+        mockFile.Setup(f => f.Extension).Returns("txt");
+        mockFile.Setup(f => f.Identifier).Returns("WOPITEST");
+        mockFile.Setup(f => f.LastWriteTimeUtc).Returns(DateTime.UtcNow);
+
+        var httpContext = new DefaultHttpContext();
+
+        var builder = new DefaultCheckFileInfoBuilder(CreatePermissionProvider().Object, new WopiHostExtensions());
+        var result = await builder.BuildAsync(mockFile.Object, httpContext);
+
+        Assert.Null(result.FileUrl);
+    }
+
+    [Fact]
+    public async Task GetWopiCheckFileInfo_CallsOnCheckFileInfoEvent()
+    {
+        var mockFile = new Mock<IWopiFile>();
+        // Stub Checksum so GetEncodedSha256 takes the early-return path; otherwise it would
+        // call OpenReadAsync (also unmocked), and ComputeHashAsync(null) throws.
+        mockFile.Setup(f => f.Checksum).Returns(new ReadOnlyMemory<byte>([0]));
+        mockFile.Setup(f => f.Name).Returns("test");
+        mockFile.Setup(f => f.Owner).Returns("owner");
+        mockFile.Setup(f => f.Extension).Returns("txt");
+        mockFile.Setup(f => f.LastWriteTimeUtc).Returns(DateTime.UtcNow);
+        mockFile.Setup(f => f.Exists).Returns(true);
+        mockFile.Setup(f => f.Length).Returns(12345);
+        var eventFired = false;
+        var extensions = new RewritingExtensions
+        {
+            CheckFileInfoHandler = (context, _) => { eventFired = true; return Task.FromResult(context.CheckFileInfo); },
+        };
+        var httpContext = new DefaultHttpContext();
+
+        var builder = new DefaultCheckFileInfoBuilder(CreatePermissionProvider().Object, extensions);
+        _ = await builder.BuildAsync(mockFile.Object, httpContext);
+
+        Assert.True(eventFired);
+    }
+
+    [Fact]
+    public async Task GetWopiCheckFileInfo_WithAuthenticatedUser()
+    {
+        var mockFile = new Mock<IWopiFile>();
+        // Stub Checksum so GetEncodedSha256 takes the early-return path; otherwise it would
+        // call OpenReadAsync (also unmocked), and ComputeHashAsync(null) throws.
+        mockFile.Setup(f => f.Checksum).Returns(new ReadOnlyMemory<byte>([0]));
+        mockFile.Setup(f => f.Name).Returns("test");
+        mockFile.Setup(f => f.Owner).Returns("owner");
+        mockFile.Setup(f => f.Extension).Returns("txt");
+        mockFile.Setup(f => f.LastWriteTimeUtc).Returns(DateTime.UtcNow);
+        mockFile.Setup(f => f.Exists).Returns(true);
+        mockFile.Setup(f => f.Length).Returns(12345);
+        var mockSecurityHandler = new Mock<IWopiPermissionProvider>();
+        mockSecurityHandler
+            .Setup(_ => _.GetFilePermissionsAsync(It.IsAny<ClaimsPrincipal>(), It.IsAny<IWopiFile>(), It.IsAny<CancellationToken>()))
+            .ReturnsAsync(WopiFilePermissions.None);
+        var httpContext = new DefaultHttpContext()
+        {
+            User = new ClaimsPrincipal(
+                new ClaimsIdentity(
+                [
+                    new(ClaimTypes.NameIdentifier, "userId"),
+                    new(ClaimTypes.Name, "test")
+                ], "test auth scheme")),
+        };
+
+        var builder = new DefaultCheckFileInfoBuilder(mockSecurityHandler.Object, new WopiHostExtensions());
+        var result = await builder.BuildAsync(mockFile.Object, httpContext);
+
+        Assert.Equal("userId", result.UserId);
+        Assert.Equal("userId", result.HostAuthenticationId);
+        Assert.Equal("test", result.UserFriendlyName);
+        Assert.Null(result.UserPrincipalName);
+        Assert.False(result.ReadOnly);
+        Assert.False(result.RestrictedWebViewOnly);
+        Assert.False(result.UserCanAttend);
+        Assert.False(result.UserCanNotWriteRelative);
+        Assert.False(result.UserCanPresent);
+        Assert.False(result.UserCanRename);
+        Assert.False(result.UserCanWrite);
+        Assert.False(result.WebEditingDisabled);
+        Assert.False(result.IsAnonymousUser);
+    }
+
+    [Fact]
+    public void BuildCheckFolderInfo_ReturnsAnonymousUser_WhenNotAuthenticated()
+    {
+        var mockFolder = new Mock<IWopiFolder>();
+        mockFolder.Setup(f => f.Name).Returns("MyFolder");
+        var httpContext = new DefaultHttpContext();
+
+        var builder = new DefaultCheckFolderInfoBuilder();
+        var result = builder.Build(mockFolder.Object, httpContext);
+
+        Assert.Equal("MyFolder", result.FolderName);
+        Assert.True(result.IsAnonymousUser);
+        Assert.Null(result.UserId);
+        Assert.Null(result.UserFriendlyName);
+    }
+
+    [Fact]
+    public void BuildCheckFolderInfo_WithAuthenticatedUser()
+    {
+        var mockFolder = new Mock<IWopiFolder>();
+        mockFolder.Setup(f => f.Name).Returns("MyFolder");
+        var httpContext = new DefaultHttpContext()
+        {
+            User = new ClaimsPrincipal(
+                new ClaimsIdentity(
+                [
+                    new(ClaimTypes.NameIdentifier, "userId42"),
+                    new(ClaimTypes.Name, "Jane Doe")
+                ], "test auth scheme")),
+        };
+
+        var builder = new DefaultCheckFolderInfoBuilder();
+        var result = builder.Build(mockFolder.Object, httpContext);
+
+        Assert.Equal("userId42", result.UserId);
+        Assert.Equal("Jane Doe", result.UserFriendlyName);
+        Assert.False(result.IsAnonymousUser);
+    }
+
+    // Note: OnCheckFolderInfo callback firing is no longer the builder's responsibility — it
+    // moved to FoldersController.CheckFolderInfo as part of #363's resolution. The callback
+    // round-trip is covered by FoldersControllerTests.CheckFolderInfo_CallsOnCheckFolderInfoEvent.
+
+    [Fact]
+    public async Task GetWopiCheckContainerInfo_CallsOnCheckContainerInfoEvent()
+    {
+        var mockFolder = new Mock<IWopiFolder>();
+        mockFolder.Setup(f => f.Name).Returns("test");
+        var mockSecurityHandler = new Mock<IWopiPermissionProvider>();
+        mockSecurityHandler
+            .Setup(_ => _.GetContainerPermissionsAsync(It.IsAny<ClaimsPrincipal>(), It.IsAny<IWopiFolder>(), It.IsAny<CancellationToken>()))
+            .ReturnsAsync(WopiContainerPermissions.None);
+        var eventFired = false;
+        var extensions = new RewritingExtensions
+        {
+            CheckContainerInfoHandler = (context, _) => { eventFired = true; return Task.FromResult(context.CheckContainerInfo); },
+        };
+        var httpContext = new DefaultHttpContext();
+
+        var builder = new DefaultCheckContainerInfoBuilder(mockSecurityHandler.Object, extensions);
+        var result = await builder.BuildAsync(mockFolder.Object, httpContext);
+
+        Assert.Equal("test", result.Name);
+        Assert.True(eventFired);
+    }
+
+    [Fact]
+    public async Task GetWopiCheckContainerInfo_WithAllPermissions_ReturnsAllTrue()
+    {
+        var mockFolder = new Mock<IWopiFolder>();
+        mockFolder.Setup(f => f.Name).Returns("MyContainer");
+        var mockSecurityHandler = new Mock<IWopiPermissionProvider>();
+        mockSecurityHandler
+            .Setup(_ => _.GetContainerPermissionsAsync(It.IsAny<ClaimsPrincipal>(), It.IsAny<IWopiFolder>(), It.IsAny<CancellationToken>()))
+            .ReturnsAsync(
+                WopiContainerPermissions.UserCanCreateChildContainer |
+                WopiContainerPermissions.UserCanCreateChildFile |
+                WopiContainerPermissions.UserCanDelete |
+                WopiContainerPermissions.UserCanRename);
+        var httpContext = new DefaultHttpContext();
+
+        var builder = new DefaultCheckContainerInfoBuilder(mockSecurityHandler.Object, new WopiHostExtensions());
+        var result = await builder.BuildAsync(mockFolder.Object, httpContext);
+
+        Assert.Equal("MyContainer", result.Name);
+        Assert.True(result.UserCanCreateChildContainer);
+        Assert.True(result.UserCanCreateChildFile);
+        Assert.True(result.UserCanDelete);
+        Assert.True(result.UserCanRename);
+        Assert.False(result.IsEduUser);
+    }
+
+    [Fact]
+    public async Task GetWopiCheckContainerInfo_WithNoPermissions_ReturnsAllFalse()
+    {
+        var mockFolder = new Mock<IWopiFolder>();
+        mockFolder.Setup(f => f.Name).Returns("RestrictedContainer");
+        var mockSecurityHandler = new Mock<IWopiPermissionProvider>();
+        mockSecurityHandler
+            .Setup(_ => _.GetContainerPermissionsAsync(It.IsAny<ClaimsPrincipal>(), It.IsAny<IWopiFolder>(), It.IsAny<CancellationToken>()))
+            .ReturnsAsync(WopiContainerPermissions.None);
+        var httpContext = new DefaultHttpContext();
+
+        var builder = new DefaultCheckContainerInfoBuilder(mockSecurityHandler.Object, new WopiHostExtensions());
+        var result = await builder.BuildAsync(mockFolder.Object, httpContext);
+
+        Assert.Equal("RestrictedContainer", result.Name);
+        Assert.False(result.UserCanCreateChildContainer);
+        Assert.False(result.UserCanCreateChildFile);
+        Assert.False(result.UserCanDelete);
+        Assert.False(result.UserCanRename);
+    }
+
+    [Fact]
+    public async Task GetWopiCheckContainerInfo_WithPartialPermissions_ReturnsCorrectFlags()
+    {
+        var mockFolder = new Mock<IWopiFolder>();
+        mockFolder.Setup(f => f.Name).Returns("PartialContainer");
+        var mockSecurityHandler = new Mock<IWopiPermissionProvider>();
+        mockSecurityHandler
+            .Setup(_ => _.GetContainerPermissionsAsync(It.IsAny<ClaimsPrincipal>(), It.IsAny<IWopiFolder>(), It.IsAny<CancellationToken>()))
+            .ReturnsAsync(WopiContainerPermissions.UserCanCreateChildFile | WopiContainerPermissions.UserCanRename);
+        var httpContext = new DefaultHttpContext();
+
+        var builder = new DefaultCheckContainerInfoBuilder(mockSecurityHandler.Object, new WopiHostExtensions());
+        var result = await builder.BuildAsync(mockFolder.Object, httpContext);
+
+        Assert.False(result.UserCanCreateChildContainer);
+        Assert.True(result.UserCanCreateChildFile);
+        Assert.False(result.UserCanDelete);
+        Assert.True(result.UserCanRename);
+    }
+
+    [Fact]
+    public async Task GetWopiCheckFileInfo_WithAllPermissions_ReturnsAllTrue()
+    {
+        var mockFile = new Mock<IWopiFile>();
+        // Stub Checksum so GetEncodedSha256 takes the early-return path; otherwise it would
+        // call OpenReadAsync (also unmocked), and ComputeHashAsync(null) throws.
+        mockFile.Setup(f => f.Checksum).Returns(new ReadOnlyMemory<byte>([0]));
+        mockFile.Setup(f => f.Name).Returns("test");
+        mockFile.Setup(f => f.Owner).Returns("owner");
+        mockFile.Setup(f => f.Extension).Returns("txt");
+        mockFile.Setup(f => f.LastWriteTimeUtc).Returns(DateTime.UtcNow);
+        mockFile.Setup(f => f.Exists).Returns(true);
+        mockFile.Setup(f => f.Length).Returns(100);
+        var mockSecurityHandler = new Mock<IWopiPermissionProvider>();
+        mockSecurityHandler
+            .Setup(_ => _.GetFilePermissionsAsync(It.IsAny<ClaimsPrincipal>(), It.IsAny<IWopiFile>(), It.IsAny<CancellationToken>()))
+            .ReturnsAsync(
+                WopiFilePermissions.ReadOnly |
+                WopiFilePermissions.RestrictedWebViewOnly |
+                WopiFilePermissions.UserCanAttend |
+                WopiFilePermissions.UserCanNotWriteRelative |
+                WopiFilePermissions.UserCanPresent |
+                WopiFilePermissions.UserCanRename |
+                WopiFilePermissions.UserCanWrite |
+                WopiFilePermissions.WebEditingDisabled);
+        var httpContext = new DefaultHttpContext()
+        {
+            User = new ClaimsPrincipal(
+                new ClaimsIdentity(
+                [
+                    new(ClaimTypes.NameIdentifier, "userId"),
+                    new(ClaimTypes.Name, "test")
+                ], "test auth scheme")),
+        };
+
+        var builder = new DefaultCheckFileInfoBuilder(mockSecurityHandler.Object, new WopiHostExtensions());
+        var result = await builder.BuildAsync(mockFile.Object, httpContext);
+
+        Assert.True(result.ReadOnly);
+        Assert.True(result.RestrictedWebViewOnly);
+        Assert.True(result.UserCanAttend);
+        Assert.True(result.UserCanNotWriteRelative);
+        Assert.True(result.UserCanPresent);
+        Assert.True(result.UserCanRename);
+        Assert.True(result.UserCanWrite);
+        Assert.True(result.WebEditingDisabled);
+    }
+
+    [Fact]
+    public async Task GetWopiCheckContainerInfo_ThrowsOnNullContainer()
+    {
+        IWopiFolder? container = null;
+        var httpContext = new DefaultHttpContext();
+
+        var builder = new DefaultCheckContainerInfoBuilder(CreatePermissionProvider().Object, new WopiHostExtensions());
+        await Assert.ThrowsAsync<ArgumentNullException>(() => builder.BuildAsync(container!, httpContext));
+    }
+
+    [Fact]
+    public async Task GetWopiCheckContainerInfo_ThrowsOnNullHttpContext()
+    {
+        var mockFolder = new Mock<IWopiFolder>();
+
+        var builder = new DefaultCheckContainerInfoBuilder(CreatePermissionProvider().Object, new WopiHostExtensions());
+        await Assert.ThrowsAsync<ArgumentNullException>(() => builder.BuildAsync(mockFolder.Object, null!));
+    }
+
+    /// <summary>
+    /// Capturing <see cref="IWopiHostExtensions"/> used by tests to observe / rewrite the
+    /// <c>CheckFileInfo</c> and <c>CheckContainerInfo</c> hook calls.
+    /// </summary>
+    private sealed class RewritingExtensions : WopiHostExtensions
+    {
+        public Func<WopiCheckFileInfoContext, CancellationToken, Task<WopiCheckFileInfo>>? CheckFileInfoHandler { get; set; }
+        public Func<WopiCheckContainerInfoContext, CancellationToken, Task<WopiCheckContainerInfo>>? CheckContainerInfoHandler { get; set; }
+
+        public override Task<WopiCheckFileInfo> OnCheckFileInfoAsync(WopiCheckFileInfoContext context, CancellationToken cancellationToken = default)
+            => CheckFileInfoHandler is null
+                ? base.OnCheckFileInfoAsync(context, cancellationToken)
+                : CheckFileInfoHandler(context, cancellationToken);
+
+        public override Task<WopiCheckContainerInfo> OnCheckContainerInfoAsync(WopiCheckContainerInfoContext context, CancellationToken cancellationToken = default)
+            => CheckContainerInfoHandler is null
+                ? base.OnCheckContainerInfoAsync(context, cancellationToken)
+                : CheckContainerInfoHandler(context, cancellationToken);
+    }
+}
